+++ addons/binding/org.openhab.binding.tradfri.test/src/test/java/org/openhab/binding/tradfri/discovery/TradfriDiscoveryParticipantOSGITest.java
@@ -16,7 +16,7 @@ import static org.hamcrest.CoreMatchers.*;
 import static org.junit.Assert.*;
 import static org.mockito.Mockito.when;
 import static org.mockito.MockitoAnnotations.initMocks;
-import static org.openhab.binding.tradfri.TradfriBindingConstants.*;
+import static org.openhab.binding.tradfri.internal.TradfriBindingConstants.*;
 
 import javax.jmdns.ServiceInfo;
