+++ addons/binding/org.openhab.binding.tradfri/src/main/java/org/openhab/binding/tradfri/internal/DeviceUpdateListener.java
@@ -12,7 +12,7 @@
  */
 package org.openhab.binding.tradfri.internal;
 
-import org.openhab.binding.tradfri.handler.TradfriGatewayHandler;
+import org.openhab.binding.tradfri.internal.handler.TradfriGatewayHandler;
 
 import com.google.gson.JsonObject;
