+++ addons/binding/org.openhab.binding.tradfri/src/main/java/org/openhab/binding/tradfri/internal/model/TradfriControllerData.java
@@ -12,7 +12,7 @@
  */
 package org.openhab.binding.tradfri.internal.model;
 
-import static org.openhab.binding.tradfri.TradfriBindingConstants.SWITCH;
+import static org.openhab.binding.tradfri.internal.TradfriBindingConstants.SWITCH;
 
 import com.google.gson.JsonElement;
